requirements.txt
Django==1.6.1
django-bootstrap3==2.5.6
django-gravatar2==1.1.3
django-user-streams==0.6.0
dulwich==0.9.4
ipython==1.2.0
lockfile==0.9.1
pycrypto==2.6.1
wsgiref==0.1.2
